fix on attack state

+++ Engine/Game/Assets/Scripts/PlayerController.h
@@ -80,7 +80,7 @@ public:
 
     Weapon* GetWeapon() const { return mWeapon; }
     Weapon* GetSpecialWeapon() const { return mSpecialWeapon; }
-    float GetCurrentEnergy() const { return mCurrentEnergy; }
+    int GetCurrentEnergy() const { return mCurrentEnergy; }
     EnergyType GetEnergyType() const { return mEnergyType; }
 
     void SetMovementSpeed(float percentage) { mPlayerSpeed *= percentage; }
